[qca-ssdk-shell]: sync backspace and move left keys
At the ssdk_sh prompt, while typing a command, when we use backspace
or move left buttons, the sync is lost.
When we use Backspace to remove entire string typed, the last character
still appears.
This change flushes the stdout buffer for sync. In handle_delete, the
last character is taken care by modifying the loop where we reprint
from delete position.

Change-Id: If36542318737206b23d3ff3fc54d07b80e752c14
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>